bucketlist/requirements.txt
coverage==4.0.3
coveralls==1.1
dj-database-url==0.4.0
Django==1.9.1
django-rest-swagger==0.3.5
djangorestframework==3.3.2
djangorestframework-jwt==1.7.2
docopt==0.6.2
Faker==0.0.4
Markdown==2.6.5
PyJWT==1.4.0
PyYAML==3.11
requests==2.9.1
wheel==0.24.0
